Remove question description from editor and put in [id]
